refactor(functions): extract filter move
